+++ mangadex_downloader/main.py
@@ -264,7 +264,6 @@ def download(
     use_alt_details=False,
     no_group_name=False,
     group=None,
-    legacy_sorting=False,
     use_chapter_title=False,
     unsafe=False,
     no_verify=False,
@@ -304,8 +303,6 @@ def download(
         If ``True``, Do not use scanlation group name for each chapter.
     group: :class:`str` (default: ``None``)
         Use different scanlation group for each chapter.
-    legacy_sorting: :class:`bool` (default: ``False``)
-        if ``True``, Enable legacy sorting chapter images for old reader application.
     use_chapter_title: :class:`bool` (default: ``False``)
         If ``True``, use chapter title for each chapters.
         NOTE: This option is useless if used with any single format.
@@ -440,7 +437,6 @@ def download(
             m,
             compressed_image,
             replace,
-            legacy_sorting,
             no_verify,
             kwargs_iter_chapter_images
         )
@@ -496,7 +492,6 @@ def download_chapter(
     compressed_image=False,
     save_as=default_save_as_format,
     no_group_name=False,
-    legacy_sorting=False,
     use_chapter_title=False,
     unsafe=False,
     no_verify=False
@@ -521,8 +516,6 @@ def download_chapter(
         Choose save as format
     no_group_name: :class:`bool` (default: ``False``)
         If ``True``, Do not use scanlation group name for each chapter.
-    legacy_sorting: :class:`bool` (default: ``False``)
-        if ``True``, Enable legacy sorting chapter images for old reader application.
     use_chapter_title: :class:`bool` (default: ``False``)
         If ``True``, use chapter title for each chapters.
         NOTE: This option is useless if used with any single format.
@@ -593,7 +586,6 @@ def download_chapter(
         manga,
         compressed_image,
         replace,
-        legacy_sorting,
         no_verify,
         kwargs_iter_chapter_images
     )
@@ -614,7 +606,6 @@ def download_list(
     save_as=default_save_as_format,
     no_group_name=False,
     group=None,
-    legacy_sorting=False,
     use_chapter_title=True,
     unsafe=False,
     no_verify=False
@@ -637,8 +628,6 @@ def download_list(
         If ``True``, Do not use scanlation group name for each chapter.
     group: :class:`str` (default: ``None``)
         Use different scanlation group for each chapter.
-    legacy_sorting: :class:`bool` (default: ``False``)
-        if ``True``, Enable legacy sorting chapter images for old reader application.
     use_chapter_title: :class:`bool` (default: ``False``)
         If ``True``, use chapter title for each chapters.
         NOTE: This option is useless if used with any single format.
@@ -667,7 +656,6 @@ def download_list(
             language=language,
             no_group_name=no_group_name,
             group=group,
-            legacy_sorting=legacy_sorting,
             use_chapter_title=use_chapter_title,
             unsafe=unsafe,
             no_verify=no_verify